Add Compiler Arguments To Validation Calls
So we can modify/disable where we want to.
