+++ src/Compiler/Validate/AllSctStarsMustHaveJoinedRoute.cs
@@ -4,6 +4,7 @@ using System.Text;
 using Compiler.Event;
 using Compiler.Model;
 using Compiler.Error;
+using Compiler.Argument;
 
 namespace Compiler.Validate
 {
@@ -16,7 +17,7 @@ namespace Compiler.Validate
             null
         );
 
-        public void Validate(SectorElementCollection sectorElements, IEventLogger events)
+        public void Validate(SectorElementCollection sectorElements, CompilerArguments args, IEventLogger events)
         {
             foreach (SidStarRoute sid in sectorElements.StarRoutes)
             {
