+++ src/Compiler/Validate/AllSidsMustBeUnique.cs
@@ -3,12 +3,13 @@ using Compiler.Event;
 using Compiler.Model;
 using Compiler.Error;
 using System.Collections.Generic;
+using Compiler.Argument;
 
 namespace Compiler.Validate
 {
     public class AllSidsMustBeUnique : IValidationRule
     {
-        public void Validate(SectorElementCollection sectorElements, IEventLogger events)
+        public void Validate(SectorElementCollection sectorElements, CompilerArguments args, IEventLogger events)
         {
             List<string> keys = new List<string>();
             foreach (SidStar sidStar in sectorElements.SidStars)
