+++ src/Compiler/Validate/IValidationRule.cs
@@ -1,10 +1,11 @@
 ﻿using Compiler.Event;
 using Compiler.Model;
+using Compiler.Argument;
 
 namespace Compiler.Validate
 {
     public interface IValidationRule
     {
-        public void Validate(SectorElementCollection sectorElements, IEventLogger events);
+        public void Validate(SectorElementCollection sectorElements, CompilerArguments args, IEventLogger events);
     }
 }
