+++ tests/CompilerTest/Validate/AllColoursMustBeValidTest.cs
@@ -4,6 +4,7 @@ using Compiler.Event;
 using Compiler.Error;
 using Compiler.Validate;
 using Moq;
+using Compiler.Argument;
 
 namespace CompilerTest.Validate
 {
@@ -17,6 +18,7 @@ namespace CompilerTest.Validate
         private readonly Colour fourth;
         private readonly Colour fifth;
         private readonly AllColoursMustBeValid rule;
+        private readonly CompilerArguments args;
 
         public AllColoursMustBeValidTest()
         {
@@ -28,6 +30,7 @@ namespace CompilerTest.Validate
             this.fourth = new Colour("colour4", 16777215, "test");
             this.fifth = new Colour("colour5", 16777216, "test");
             this.rule = new AllColoursMustBeValid();
+            this.args = new CompilerArguments();
         }
 
         [Fact]
@@ -36,7 +39,7 @@ namespace CompilerTest.Validate
             this.sectorElements.Add(this.second);
             this.sectorElements.Add(this.third);
             this.sectorElements.Add(this.fourth);
-            this.rule.Validate(this.sectorElements, this.loggerMock.Object);
+            this.rule.Validate(this.sectorElements, this.args, this.loggerMock.Object);
             this.loggerMock.Verify(foo => foo.AddEvent(It.IsAny<ValidationRuleFailure>()), Times.Never);
         }
 
@@ -44,7 +47,7 @@ namespace CompilerTest.Validate
         public void TestItFailsOnNegativeValues()
         {
             this.sectorElements.Add(this.first);
-            this.rule.Validate(this.sectorElements, this.loggerMock.Object);
+            this.rule.Validate(this.sectorElements, this.args, this.loggerMock.Object);
             this.loggerMock.Verify(foo => foo.AddEvent(It.IsAny<ValidationRuleFailure>()), Times.Once);
         }
 
@@ -52,7 +55,7 @@ namespace CompilerTest.Validate
         public void TestItFailsOnValuesOverMaximum()
         {
             this.sectorElements.Add(this.fifth);
-            this.rule.Validate(this.sectorElements, this.loggerMock.Object);
+            this.rule.Validate(this.sectorElements, this.args, this.loggerMock.Object);
             this.loggerMock.Verify(foo => foo.AddEvent(It.IsAny<ValidationRuleFailure>()), Times.Once);
         }
     }
