+++ tests/CompilerTest/Validate/AllColoursMustHaveAUniqueIdTest.cs
@@ -4,6 +4,7 @@ using Compiler.Event;
 using Compiler.Error;
 using Compiler.Validate;
 using Moq;
+using Compiler.Argument;
 
 namespace CompilerTest.Validate
 {
@@ -15,6 +16,7 @@ namespace CompilerTest.Validate
         private readonly Colour second;
         private readonly Colour third;
         private readonly AllColoursMustHaveAUniqueId rule;
+        private readonly CompilerArguments args;
 
         public AllColoursMustHaveAUniqueIdTest()
         {
@@ -24,6 +26,7 @@ namespace CompilerTest.Validate
             this.second = new Colour("colour2", 123, "test");
             this.third = new Colour("colour1", 123, "test");
             this.rule = new AllColoursMustHaveAUniqueId();
+            this.args = new CompilerArguments();
         }
 
         [Fact]
@@ -31,7 +34,7 @@ namespace CompilerTest.Validate
         {
             this.sectorElements.Add(this.first);
             this.sectorElements.Add(this.second);
-            this.rule.Validate(sectorElements, this.loggerMock.Object);
+            this.rule.Validate(sectorElements, this.args, this.loggerMock.Object);
 
             this.loggerMock.Verify(foo => foo.AddEvent(It.IsAny<ValidationRuleFailure>()), Times.Never);
         }
@@ -42,7 +45,7 @@ namespace CompilerTest.Validate
             this.sectorElements.Add(this.first);
             this.sectorElements.Add(this.second);
             this.sectorElements.Add(this.third);
-            this.rule.Validate(sectorElements, this.loggerMock.Object);
+            this.rule.Validate(sectorElements, this.args, this.loggerMock.Object);
             this.loggerMock.Verify(foo => foo.AddEvent(It.IsAny<ValidationRuleFailure>()), Times.Once);
         }
     }
